refactor: 💡 unique slice as app context
